CAMS-87: Rename cosmos humble to cosmos client humble and rename files
